Initial Refacotring of Backtest
